requirements.txt
torch
torchvision
ultralytics
opencv-python
numpy
faiss-cpu
httpx
ffmpeg-python
pyyaml
scipy
